lastrelease

+++ ft_bzero.c
@@ -3,14 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_bzero.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tombanks <tombanks@student.42.fr>          +#+  +:+       +#+        */
+/*   By: tcharvet <tcharvet@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/28 20:41:02 by tcharvet          #+#    #+#             */
-/*   Updated: 2020/11/28 20:41:40 by tombanks         ###   ########.fr       */
+/*   Updated: 2021/01/19 13:17:53 by tcharvet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	ft_bzero(void *s, unsigned int n)
+#include <string.h>
+
+void	ft_bzero(void *s, size_t n)
 {
 	while (n-- > 0)
 		*(char*)s++ = 0;
